Issue #168: Changes to compile_program.sh
To prevent the accumulation of the compilation directories, an argument  was added to compile_program.sh to determine whether or not this directory should be removed after a succesful compilation of a program. The default behavior is DELETE.

In order to NOT delete this directory, the user must follow the following example:

compile_program.sh program.f90 full no

+++ src/programs/compile_program.sh
@@ -7,6 +7,7 @@ if [ $# -lt 1 ]; then
 fi
 program=$(basename $1 .f90)
 mode=$2
+deleteCompileDir=$3
 
 export COMPILE_MIDAS_ADD_DEBUG_OPTIONS=${COMPILE_MIDAS_ADD_DEBUG_OPTIONS:-no}
 . ./commons/compile_setup.sh
@@ -32,6 +33,21 @@ else
   echo "..."
   echo "... > Compiling only routine : $mode"
 fi
+
+if [ "$deleteCompileDir" == "" ] ; then
+  echo "..."
+  echo "... !WARNING! no directive given for the deletion of the compilation directory, assuming 'yes'"
+  deleteCompileDir=yes
+fi
+
+if [ $deleteCompileDir == yes ] ; then
+  echo "..."
+  echo "... > the compilation directory will be DELETED after the successful completion of this task"
+else
+  echo "..."
+  echo "... > the compilation directory will be KEPT"
+fi
+
 
 echo "..."
 echo "... > Compiling for an MPI executable"
@@ -153,7 +169,9 @@ if [ $mode == full ] ; then
   fi
   cp ${midasAbs} ${absdir}/
 
-  rm -rf ${compiledir}
+  if [ $deleteCompileDir == yes ] ; then
+      rm -rf ${compiledir}
+  fi
 
 elif [ $mode == abs ] ; then
